Introduce `observeQuick`
